+++ backend/src/authentication/mod.rs
@@ -1,5 +1,7 @@
 //! backend/src/authentication/mod.rs
 //! Holds logic and helpers related to authenticating users.
+pub mod middleware;
 pub mod password;
 
+pub use middleware::*;
 pub use password::*;