+++ backend/src/configuration.rs
@@ -1,8 +1,8 @@
 //! backend/src/configuration.rs
 //! Logic to read configuration files and create structs to be used
 //! throughout the rest of the application.
-use crate::surrealdb_repo::Database;
 use config::ConfigError;
+use secrecy::Secret;
 use serde::Deserialize;
 use serde_aux::field_attributes::deserialize_number_from_string;
 use std::path::PathBuf;
@@ -49,6 +49,7 @@ pub struct ApplicationSettings {
     pub port: u16,
     pub host: String,
     pub base_url: String,
+    pub hmac_secret: Secret<String>,
 }
 
 /// Struct to hold information regarding the database